refactor(models/location_picker_store): :recycle: Renamed location_picker_store to create_location_page_store
